+++ packages/werkbank/lib/src/addons/src/state/src/mutable_state_id.dart
@@ -1,0 +1,2 @@
+/// {@category Keeping State}
+extension type MutableStateId(String _id) {}